+++ client/components/GlobalControls/globalControls.scss
@@ -1,0 +1,12 @@
+.global-controls-component {
+	display: flex;
+	align-items: center;
+
+	.bp3-button {
+		padding: 0px 10px;
+		color: currentColor;
+		.bp3-icon {
+			color: currentColor;
+		}
+	}
+}
