+++ client/components/GlobalControls/index.ts
@@ -1,0 +1,1 @@
+export { default } from './GlobalControls';
