+++ client/components/Header/Header.tsx
@@ -1,13 +1,11 @@
-import React, { useState } from 'react';
+import React from 'react';
 import classNames from 'classnames';
-import { Button, AnchorButton, Intent } from '@blueprintjs/core';
+import { AnchorButton, Intent } from '@blueprintjs/core';
 
-import { GridWrapper, HeaderControls } from 'components';
+import { CommunityHeroButton } from 'types';
+import { GridWrapper, GlobalControls } from 'components';
 import { usePageContext } from 'utils/hooks';
 import { getResizedUrl } from 'utils/images';
-import { apiFetch } from 'client/utils/apiFetch';
-import { CommunityHeroButton } from 'types';
-import { useViewport } from 'client/utils/useViewport';
 
 require('./header.scss');
 
@@ -22,29 +20,7 @@ const defaultProps = {
 type Props = OwnProps & typeof defaultProps;
 
 const Header = (props: Props) => {
-	const { locationData, communityData, loginData, scopeData } = usePageContext(
-		props.previewContext,
-	);
-	const [isLoading, setIsLoading] = useState(false);
-	const handleLogout = () => {
-		apiFetch('/api/logout').then(() => {
-			window.location.href = '/';
-		});
-	};
-	const { viewportWidth } = useViewport();
-	const isMobile = viewportWidth! <= 750;
-	const handleCreatePub = () => {
-		setIsLoading(true);
-		return apiFetch
-			.post('/api/pubs', { communityId: communityData.id })
-			.then((newPub) => {
-				window.location.href = `/pub/${newPub.slug}`;
-			})
-			.catch((err) => {
-				console.error(err);
-				setIsLoading(false);
-			});
-	};
+	const { locationData, communityData, loginData } = usePageContext(props.previewContext);
 
 	const calculateComponentClasses = (hideHero) => {
 		let dynamicComponentClasses = '';
@@ -184,7 +160,7 @@ const Header = (props: Props) => {
 							</a>
 						)}
 					</div>
-					<HeaderControls isBasePubPub={locationData.isBasePubPub} loggedIn={loggedIn} />
+					<GlobalControls isBasePubPub={locationData.isBasePubPub} loggedIn={loggedIn} />
 				</GridWrapper>
 			</div>
 			{!hideHero && (
