+++ client/components/Header/header.scss
@@ -45,20 +45,6 @@
 					display: block;
 				}
 			}
-			.buttons-wrapper {
-				display: flex;
-				align-items: center;
-				position: relative;
-				left: 10px;
-				.header-dashboard-button {
-					.bp3-icon-caret-down {
-						color: inherit;
-					}
-				}
-				.bp3-button {
-					padding: 0px 10px;
-				}
-			}
 		}
 	}
 	.hero {
