--- client/components/HeaderControls/headerControls.scss
@@ -1,16 +1,0 @@
-.header-controls-component {
-    color: white;
-
-    &.bp3-dark {
-        a {
-            color: #fff;
-        }
-    }
-
-    .bp3-button {
-        color: currentColor;
-        .bp3-icon {
-            color: currentColor;
-        }
-    }
-}
