--- client/components/HeaderControls/index.ts
@@ -1,1 +1,0 @@
-export { default } from './HeaderControls';
