--- client/components/HeaderControls/mobileAwareButton.scss
@@ -1,15 +1,0 @@
-@import 'styles/mixins.scss';
-
-$component-prefix: 'mobile-aware-button-component';
-
-@include desktop {
-    .#{$component-prefix}__mobile-button {
-        display: none !important;
-    }
-}
-
-@include mobile {
-    .#{$component-prefix}__desktop-button {
-        display: none !important;
-    }
-}
