+++ client/components/MobileAware/index.ts
@@ -1,0 +1,1 @@
+export { default, RenderProps as MobileAwareRenderProps } from './MobileAware';
