+++ client/styles/mixins.scss
@@ -1,3 +1,5 @@
+@import './variables.scss';
+
 @mixin bleed-top-border() {
 	background-color: #fff;
 	padding: 0 8px;
@@ -6,14 +8,14 @@
 	transform: translateY(-50%);
 }
 
-@mixin mobile($mobile-viewport-cutoff: 750px) {
-	@media screen and (max-width: $mobile-viewport-cutoff) {
+@mixin mobile($cutoff: $mobile-cutoff) {
+	@media screen and (max-width: $cutoff) {
 		@content;
 	}
 }
 
-@mixin desktop($mobile-viewport-cutoff: 750px) {
-	@media screen and (min-width: #{1 + $mobile-viewport-cutoff}) {
+@mixin desktop($cutoff: $mobile-cutoff) {
+	@media screen and (min-width: #{1 + $cutoff}) {
 		@content;
 	}
 }
